GSM code update
M1 transitions were not given in units of magnetons in EM transitions. It is corrected. M1 cross sections were correct.
The units of spherical harmonics was not standard. It has been corrected. It was not visible in previous applications.
An option involving the KKNN potential and HF calculations was wrongly handled. This option had never been used before. It has been corrected.
Allocation and deallocation of arrays, arrays of arrays, etc, has been automatized.
Hypernuclei are being coded, but the code has not been tested with hypernuclei. Hence, no information is being given about it.