GSM and GSM-CC upgrade
Inclusion of many-body projectiles and no-core framework in GSM-CC are in development.
EM transition strengths in GSM-CC were not correct. This option had never been used before. It is corrected.
Rms radii functions were not correct with standard HO-SM. This option had never been used before. It is corrected.
Errors found and corrected in GSM multipoles. This option had never been used before.
Errors found in the GSM and GSM-CC hybrid 1D/2D codes in proton-neutron configuration print in GSM, use of MSDHF in GSM, observables in GSM-CC. They are corrected.
Hybrid 1D/2D codes are very rarely used.
Error found in GSM-2D with proton-neutron one-jumps put to full or partial storage and Hamiltonian to on-the-fly. It is corrected. The bug was invisible as previous options had not been considered in applications.