Few minor problems found
A few input files in exercises did not run as options were not written correctly. It was simple to correct anyway with messages on screen. It is corrected.
A few codes almost never used (GSM_H_eigenvectors_dir,  GSM_HF_OCM_basis_only_dir, GSM_OBMEs_TBMEs_dir,  GSM_observables_dir) did not compile. It is corrected. They are not used in exercises and are not mentioned in the manual.