Bug in a rarely used routine and several changes
A bug was found in a+ a~ matrix elements : some values were missing in the output files. 
a+[p].a~[p] was not appearing on screen and was written "?" instead.
It has been corrected.

Comments have been added in GSM-CC and a few comments have been corrected in GSM.
A few technical changes have been done in MPI transfer due to changing standards. They are invisible to users.